ANN Test update
currently testing to modulate a sine Function depending on the currently
activated neuron in the input layer

+++ KNN(StraightForward)/mainwindow.h
@@ -43,7 +43,7 @@ private:
     unsigned int mSize,mLayers;
 
     QGraphicsScene *errorGraph;
-    unsigned int mFrameCounter;
+    unsigned long int mFrameCounter;
     float lastError,accError;
 
     QGraphicsScene *outPlot;
